requirements.txt
discord
discord.py
Pillow
requests
Pyrebase4
discord-py-interactions
